Simple "pretty-printing" logic for writing out AST.
This isn't really right, and the approach might be too simplistic to work in the end, but I'm bypassing the whole Spire back-end for now and just directly writing out the AST again as HLSL. Considering that the code came in as HLSL, the hope is that there won't be too far to go in getting something working out on the other side.

This currently gets my first test HLSL shader through Spire and back to HLSL which passes compilation with the `D3DCompile` API.
I don't yet know whether the HLSL it produced is actually something that retains the semantics of the original code, but it looks good visually (I expect any differences will be in any binding/semantic meta-data I might be dropping).

A few notes on details of the pretty-printing logic:

- I need to re-generate `cbuffer` declarations from the desugared form I added in previous checkins. Right now this is easy enough because it blindly applies to any variable using the `ConstantBuffer<T>` type. Also, when a member reference has a base expression that references such a variable, we drop the base expression.

- Because my changes have made initializers/constructors explicit, these show up like member references (looking up a member in a type or extension). I special-case references to initializers so that they refer to the type instead.

- When outputting a reference to a declaration, we need to check if it had any generic arguments applied, and if so we should print them too. I'm not being clever and trying to "re-sugar" types so that I can print `Texture2D` instead of `Texture2D<float4>`.

- Okay, I actually print `Texture2D<vector<float,4> >` which is way worse. The printing for vector types should probalby be improved, but it may also be worthwhile/important to try to preserve type-name sugar from the user's program as much as possible.

- Finally, I added a hack to the parser that identifies declarations coming from our "stdlib", so that I don't try to emit them when pretty-printing.