Clean up handling of non-proper/-usable types.
A "proper" type in type theory is one that can classify values. `std::vector<int>` is a proper type, but a type like `std::vector` is not, because you can't actually have a value of that type.

Previously we didn't really have any cases of non-proper types, so we could get away with just spot checking for the more narrow issue that `void` isn't allowed as a type in various places (although there is no semantic reason why it shouldn't be).

HLSL throws an additional wrinkle into things because a non-proper type like `Texture2D` can be used where a proper type is expected, and it automatically gets applied to default arguments to make `Texture2D<float4>`.
We already have code to do that conversion, but it isn't being applied uniformly.

This change adds specific functions `CheckProperType` and `CheckUsableType` that both check a type expression and ensure that it is appropriate for a context where either proper types or "usable" types (proper types other than `void`) are expected. It also surveys all the call sites to `TranslateTypeNode` and makes them use one or the other of the new routines.

Cast expressions used to just store the target type as an `ExpressionSyntaxNode`, but now they use a `TypeExp` like other cases in the AST, since the new routines expect a `TypeExp`.