Allow global variables of bindable resource types.
We have an error condition on `Variable` declarations with a resource type, which triggers on both globals and locals now.
I updated the test to look at the parent of the declaration, so that it only gives an error when the parent is a `ScopeDecl`, which is the pseudo-declaration we use for block statements (which should be the parent of any local variables).