Add `SwizzleExpr`
The code had previous just left vector swizzles as `MemberExpressionSyntaxNode`, which then requires downstream code to keep figuring out that it actually *isn't* a member reference, but a swizzle.

This change adds a distinct AST node to be produced by semantic checking. It also ends up re-writing the logic for checking swizzles, but should produce basically the same results.