Add rudimentary generic constraint solver to overload resolution.
The basic idea is that if you are trying to apply something that is generic, you create a bunch of fresh type-level variables to apply the generic to, and then during checking of arguments against parameter types you generate a set of constraints needed to infer their types.

This is currently only being applied in the narrow case of aggregate type constructors that are found via extensions, but that is mostly due to limitations in how I'm setting up overloads, rather than a fundamental issue.