Allow invalid preprocessor directives in skipped code.
I changed the structure a bit so that invalid directives get mapped to a handler like everything else (to avoid having to include an extra null-pointer check), and that hanlder doesn't have the flag that forces it to be processed when in skipping mode.

This helps with code that tries to cross-compile as either HLSL or GLSL, and that `#ifdef`s out code that uses GLSL-specific directives like `#extension`.

(Yes, a more complete solution would be to handle `#extension` directives, and I guess I should get to that eventually)