Some cleanups to overload resolution.
This change allows overload resolution to actually track a list of all the candidates under consideration (in the case where it can't determine which is best). This is helpful even when there will be a final best candidate, since we need to compare it to the candidates currently under consideration to be sure it is the best one.