Single-component swizzles yield scalars.
This is an obvious fix that I probably should have caught when implementing swizzles.

Note that HLSL actually claims that single-component swizzles yield 1-element vectors (and then allows 1-element vectors to implicitly convert to scalars). This seems silly, though, so I'm going to avoid the whole notion of 1-element vectors (and 1x1, Nx1 and 1xM matrices) for as long as possible.