Fix parsing of numeric literals starting with `-.`
The parser code already had a problem that it would parse `.8` as two tokens: `.` and `8`.
A naive fix for that would be to switch into `State::Int` when the lookahead is `.` and a digit.

Unfortunately, that wouldn't handle `-.5` because in that case the `-` makes us go into the `State::Operator`, which would consume teh `-.` and then break it apart, leaving us with three tokens: `-`, `.`, and `5`.

The only possible fix is to not treat `.` as an operator character by default, and to special-case it a bit. Right now this means we can't parse `...` properly (but shading languages don't use that token right now).

This is all kind of a mess, and I really just want to rewrite the lexer.