Very basic support for function redeclaration.
This currently just allows functions to be redeclared (but not redefined), but doesn't actually do any work for later steps (like overload resolution) to properly treat the set of declarations as a single candidate.

Also, longer term we are going to need to ask ourselves how rate- or target-specific functions should integrate into this model. Should they be considered as distinct overloads, disambiguated by overload resolution, or as distinct specialized definition for the same function, so that overload resolution picks the right *set* of definitions, but then target-specific code generation further refines this.

(I prefer the latter, but it would be the trickier one to implement)