Beef up handling of statement nesting.
This allows a `break` statement to find either a `switch` or a loop, and also adds support for `case` and `default` to check for an enclosing `switch`.

This also makes the "child" statement in each of these cases store a pointer to its "parent" statement, so that later stages don't have to re-discover the control flow relationship if they don't want to.