Add "profile" support, and test tool functionality.
This is mostly to support adding HLSL tests.

When I say "profile" I mean what the standard HLSL front-end expresses as things like `vs_5_0`.
A profile combines a target stage (`vs` for the vertex stage), along with a version that specifies capabilities that should be allowed.

For our purposes things are a bit more complicated, because we have to deal with:

- Both an input and an output language, which might have different notions of versions/profiles
  - We might eventually want to support HLSL, GLSL (GL, ES, and VK flavors), DXBC (+assembly), DXIL(+assembly), and SPIRV(GL and VK flavors)
- The possibility of zero, one, or more entry points per file (which might need to be compiled for different stages, if not profiles)

Getting this right so that the compiler driver (and API) has the right "do what I mean" behavior is going to take a couple of iterations, at least.

For the beefed-up test runner we have the following changes:

- The test runner now supports basic command-line arguments
  - `-v` for verbose mode (print out command before running them, to make it easier to debug failures)
  - `-generate-hlsl-baselines` for a feature I'll describe below
  - A path prefix in order to filter tests to just those sharing that prefix

- The first line of a test case can start with `\\SPIRE_TEST_OPTS:` to indicate that the rest of the line represents options that should be added to the command line when invoking Spire on the file.
  - In particular, this can be used to specify profile and entry point names when compiling HLSL code

- Finally, there is hacky support for generating the `*.expected` files for HLSL tests using the `D3DCompile()` API.
  - A large part of why this feels hacky is that it duplicates logic that the Spire compiler already supports (for parsing Spire command-line options and turning them into `D3DCompile()` arguments. This might be easier if the spire compiler driver supported a pass-through mode where it just invoked `D3DCompile()` without actually parsing anything.

The basic result is that now we can have some test cases where we compare DX bytecode assembly output between Spire and fxc.