Fix parsing of initializers on generic value params.
The problem here is that a generic value parameter's initializer is an expression, and so it can accidentally consume the `>` that should close the generic parameter list:

    __generic<let N : int = 4> ...

We can accidentally parse the `>` as part of an expression `4 > ...` and then the loop for the generic parameter list will potentially go on looking for the closing `>` forever.

We already handle this case for generic *argument* lists by incrementing/decrementing the `genericDepth` on the parser, and then special-casing code so that operators that start with `>` are ignored as infix operators when the depth is greater than zero.

This change just makes us increment/decrement around parameter lists too.