Fix up test runner for HLSL.
Different versions of the HLSL compiler (which may all be delivered as `d3dcompiler_47.dll`) produce different output given the same input.
This means we can't check in generated baselines for HLSL tests and expect them to work across different developers' machines.

With this change, I've made the test tool auto-generate the baselines for HLSL tests if it doesn't already find them.

We should no longer check in `*.expected` files for HLSL tests.