Add HLSL test shaders from DX SDK.
I'm not sure I got everything, but this is supposed to be all the `.hlsl` and `.fx` files from the DirectX SDK (back when that was a thing).

I didn't add annotions to these files to let the test runner compile them, so this checking is hopefully clean copies of the original code.