Add a "pass-through" mode where Spire just invokes fxc.
This allows you to pass command-line options in the way that Spire expects, but then invoke the Microsoft HLSL front-end to do the actual compilation.
This is currently most useful for the test runner, where it wants to generate baselines that use fxc instead, but only have to specify the command line to use once.

I'd like to extend this at some point to also be able to invoke glslang.