Add parsing for HLSL `[attributes]`
This specifically adds the case where a square-bracket attribute has arguments, e.g.:

    [numthread(32,1,1)]

In order to not distribute existing code, I left the existing Spire style of attribute alone, and added a new kind of modifier to the AST to handle the case above.
Eventually we should consider how to streamline our AST a bit so that Spire constructs re-use the existing syntax and AST nodes a bit more.