Work on generalizing integer constant expressions.
In the context of generics, we need to support "compile-time integers" that don't have a known fixed value.
This means that types that usually store a constant value for their # of elements need to instead store an `IntVal`
which might be either a `ConstantIntVal` with a known value, or something else that is known to be a compile-time constant, but whose exact value isn't known.

This change doesn't include any other cases for `IntVal`, but the obvious one to add next is a reference to particular generic value parameter.