Change how generic parameters are inferred during overloading.
The place where this arises is when you have a generic declaration of a function/constructor/whatever, and then try to apply it to (value) arguments, without first using `<...>` to apply it to type arguments.

The first implementation I wrote tried to be "pure" (and maybe a bit too clever) by first applying the generic to "fresh" type-level variables, and then apply the resulting specialized declaration to the arguments. During ordinary checking, we'd then try to coerce argument values to parameter types, and wherever those parameter types made use of the "fresh" variables, we'd start tracking constraints on them. After ordinary checking, we'd try to solve the resulting constraint system, and only keep the candidate for overload resolution if a solution could be found.

That approach has an appealing simplicity, because it seems like we can have the type-coercion logic do double duty and also find out generic parameter constraints.
The problem that arises is that once we've solved the generic parameter constraints, we really need to go back through conversion logic a second time, so that we can actually generate the conversion operations (e.g., insert casts) and find the concrete conversion costs involved.

The above approach is more common to functional languages. The more typical approach in imperative languages (C++ and C# both do this) is to first solve for the generic arguments by trying to unify argument types with parameter types (which may reference the generic parameters), apply the generic to those arguments, and then proceed to operate on the overload candidate as before. This creates a clear separation, and ensures that we properly measure costs (and insert conversions) for arguments.

This also has the side benefit of reducing duplication with the case where we want to determine if a generic extension is applicable to a type, so that's nice.