"Fix" scoping for `for` loops.
HLSL by default has the bad kind of scoping for `for` loops:

   for(int x = 0; ...) { ... }
   int a = x; // this is okay!?!?!?

For compatibility, I'm breaking our for loop so it works the same way for now.
We can and should revisit this when we are ready to start adding language flavors/personalities to the compiler, so that we only support bogus HLSL constructs in some kind of compatibility mode.