Fix insidious single-byte bug in emit logic.
I was emitting `&=` wherever I was supposed to emit the `&` operator, which of course makes a *lot* of things go wrong.

Luckily this happened to work in a way that screwed up some static analysis from fxc, so I actually got a compile-time error instead of just inscrutable wrong results.