When in doubt, parse an expression statement.
The statement parsing logic fell through to a case where it called `Unexpected()`, but that was getting triggered whenever there was an operator at the start of the line.
The alternatives are to either list every terminal in the "first set" for expressions in the statement parsing logic, or (the much simpler option) just treat expression statements as a catch-all when the other cases don't match.