Don't even bother parsing components.
Their AST representation is deeply broken, and adds complexity in a bunch of other places.
I'm going to go ahead and remove support for now so that we can focus on getting a better baseline for variable and function declarations in place, and then we can go and add components back as specializations of those when it makes sense.