Silence a bunch of warnings.
I admit that I wasn't being at all dilligent about keeping the code clean and warning-free in my recent bunch of changes.
I don't know whether it makes sense to turn on warnings-as-errors in order to keep ourselves honest, or not.

Story Time
----------

The most annoying thing I ran into was when writing a loop like this:

    for( auto x : something ) {
        return x;
    }

This is a pretty simple idiomatic way to say "if the container has anything in it, return the first thing" and there really isn't anything a lot cleaner.
Unfortunately it causes an "unreachable code" warning because this desugars into an ordinary `for` loop with an iterator like:

    for( auto iter = something.begin() ...; ...; ++iter) {
        return x;
    }

And now that `++iter` operation is unreachable code. The only way to get rid of the warning is to manually desugar the range-based `for` loop, which seems really silly (luckily in my case the actual code simplified down to something reasonable).