Make `?:` an ordinary operator function.
This removes the need to have a special-case AST node and handling for this operator.
It can now be treated as just another ordinary overload-able operator, so that we can have both the scalar and vector versions of it (like HLSL supports).

There is a bit of a song and dance required in the parser, to recognize the two tokens `?` and `:` after the `operator` keyword and magically turn them into a single name.

Overload resolution for calls to the operator rely on the ability to compute a "join" between two types as part of generic argument inteferece. We already need similar logic for ordinary expressions, so this isn't too bad.

One big detail to note is that in HLSL the `?:` operator, as well as `&&` and `||` do *not* have the usual short-circuiting behavior that people expect.
This makes it reasonable to define them as ordinary functions.

In the long run, if we decide to have short-circuiting (e.g., for GLSL support), then I'd still advocate for having most of the semantic-checking infrastructure treat these as ordinary function calls, and only deal with the fact that they aren't in the context of code generation.