Add a notion of a type "join" for generic argument inference.
Basically, if you have a generic like:

    __generic<T> T foo(T left, T right);

and you call it with:

    int a;
    float b;
    foo(a,b);

The compiler will need to infer the generic argument for `T` from the operands.
The Right Way to do this in the context of HLSL builtins is to pick the "best" of all the types involved (in this case that is `float`).

This should really only apply to builtin types, and it should probalby also only apply to builtin generics (which are the only kind we have right now), since this isn't really an ideal feature in the general case.