Don't re-check arguments to an operator expression.
At the point where the checking logic for operator expressions delegated to the ordinary `InvokeExpression` case, it had already checked the operands of the expression, so when it called `VisitInvokeExpression` it ended up causing expressions to be re-checked that already had types and everything.
With the way the front-end currently works, there really isn't a guarantee that checking a piece of syntax again will do anything good, and we should be avoiding it anyway, since this would be an easy way to become accidentally exponential (if I check my arguments twice, and one of my arguments is a function that checks its arguments twice...).'

The fix here is simple: split the checking of invoke expressions into a preamble that checks the operands, and then a part that does the actual dispatch logic, and then have the operator case only call the latter.

(There are definitely other cleanups we can do now that operators are merged with ordinary calls.)