Implement `#line` support and make stdlib use it.
This allows us to print reasonable line numbers for errors pointing back at the stdlib code.

Even if that might not be what an end user will want to see in the long run, being able to double-click on errors messages in the Visual Studio output window and fix issues in the stdlib is a real time-saver.