Comment out some old builtin decls.
Now that things are a bit less broken in the front end, we are starting to get actual ambiguous overload errors because we have competing declarations of many builtins.
I'm commenting out a few of the old-style ones to remove spurious errors before I dive into the deeper error messages cuased by the generic declarations.