Extend type "join" to be able to deal with vectors
The basic rules are (for scalar types T and S):

    Join(vector<T,N>, vector<S,N>) -> vector<Join(T,S), N>

    Join(vector<T,N>, S) -> vector<Join(T,S), N>

This allows us to infer a correct generic argument when calling a generic with a mixture of scalars and vectors, and potentially even cases where you call it with vectors of different element types.

I don't yet deal with the case where the vector sizes might differ, but that is probably a Good Thing.

Note: There is obviously a lot of duplication between what `TryJoinTypes` and `TryUnifyTypes` do. Eventually it would probably be good to unify that logic, so that the "join" operation can add additional constraints to be solved, and handle more complex scenarios.

Note: It is probably also notable that the join/unify logic duplicates a lot of what is available via implicit coercions, which in turn duplicates a lot of what is declared as constructors. I'd like to streamline and unify this all at some point, possibly using constructors marked as usable for implicit conversion.