Add ability to declare trait conformances, and use them to filter candidates.
The basic idea is that a generic with a bound:

    __generic<T : SomeTrait> void foo(T val);

Stores both a generic parameter declaration for `T` and also a constraint `T : SomeTrait`.
When inferring arguments for the generic, at a call site like:

    int bar;
    foo(bar);

The inference system would then construct multiple constraints:

    T : int
    T : SomeTrait

And then determine that the solution `T=int` can satisfy all the constraints.