Add HLSL verification to SpireCompiler.exe
When using SpireCompiler.exe to generate HLSL, we will now attempt to load D3D compiler (D3DCompile2 function from D3DCompiler_47.dll) to verify the output HLSL and print the HLSL error output. If the required dll is not found, an error will be printed on screen and the verificaton step is skipped.